requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
psycopg2-binary
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
